done with the owner to borrow relationship

+++ all-notes/rust/flex/src/flex_usage.rs
@@ -15,7 +15,7 @@ pub fn try_flex() {
     println!("Number of bytes of flexible vector:{}",bytes(&flex_shoes.internal));
 
     let re_rep:&Vec<u8> = flex_shoes.get_internal();
-    let test_data:i32 = flex_shoes.get_data::<i32>(11);
+    let test_data:i64 = flex_shoes.get_data::<i64>(11);
 
     println!("Test data: {test_data}");
     println!("Test re-rep: {:?}",re_rep);
@@ -38,5 +38,5 @@ pub fn try_flex() {
     let stats:Vec<i64> = vec![23,77,88,99,10,11,11,11,1,1,1,99,9,999,11,22,33,22,23,33,33];
     let flex_stats:Flex = Flex::new(stats);
     println!("Number of bytes of flexible vector:{}",bytes(&flex_stats.internal));
-    println!("at 0: {}",flex_stats.get_data::<i32>(0));
+    println!("at 0: {}",flex_stats.get_data::<i64>(0));
 }
